Changing Long to Integer

+++ PuppyWorld/src/main/java/com/example/ad/repo/ServiceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.ad.domain.Services;
 
 
-public interface ServiceRepository extends JpaRepository<Services, Long> {
+public interface ServiceRepository extends JpaRepository<Services, Integer> {
 
 }
